requirements.txt
# Core
pyyaml
tqdm

# Dataset management
beir

# NLP
nltk

# Information retrieval
rank_bm25

# Embeddings
sentence-transformers
torch

# Scientific computing
numpy
scipy
scikit-learn
scikit-learn